+++ realm/realm-library/src/main/cpp/io_realm_internal_objectstore_OsAsyncOpenTask.cpp
@@ -17,10 +17,12 @@
 #include "io_realm_internal_objectstore_OsAsyncOpenTask.h"
 
 #include "util.hpp"
+#include "thread_safe_reference.hpp"
 #include "jni_util/java_method.hpp"
 #include "jni_util/java_class.hpp"
 #include "jni_util/jni_utils.hpp"
 #include "object-store/src/sync/async_open_task.hpp"
+#include "object-store/src/sync/sync_config.hpp"
 
 #include <shared_realm.hpp>
 #include <sync/async_open_task.hpp>
@@ -35,26 +37,28 @@ JNIEXPORT jlong JNICALL Java_io_realm_internal_objectstore_OsAsyncOpenTask_start
     try {
 
         static JavaClass java_async_open_task_class(env, "io/realm/internal/objectstore/OsAsyncOpenTask");
-        static JavaMethod java_notify_realm_ready(env, java_async_open_task_class, "notifyRealmReady", "()V", true);
-        static JavaMethod java_notify_error(env, java_async_open_task_class, "notifyError", "(ILjava/lang/String;)V", true);
+        static JavaMethod java_notify_realm_ready(env, java_async_open_task_class, "notifyRealmReady", "()V", false);
+        static JavaMethod java_notify_error(env, java_async_open_task_class, "notifyError", "(Ljava/lang/String;)V", false);
 
+        auto global_obj = env->NewGlobalRef(obj);
         auto& config = *reinterpret_cast<Realm::Config*>(config_ptr);
+
         std::shared_ptr<realm::AsyncOpenTask> task = Realm::get_synchronized_realm(config);
-        task->start([task_obj=obj](std::shared_ptr<Realm> realm, std::exception_ptr error) {
-        JNIEnv* local_env = jni_util::JniUtils::get_env(true);
+        task->start([task_obj=global_obj](realm::ThreadSafeReference<realm::Realm>, std::exception_ptr error) {
+            JNIEnv* local_env = jni_util::JniUtils::get_env(true);
             if (error) {
                 try {
                     std::rethrow_exception(error);
                 } catch (const std::exception& e) {
                     jstring j_error_msg = to_jstring(local_env, e.what());
-                    local_env->CallObjectMethod(java_async_open_task_class, java_notify_error, j_error_msg);
+                    local_env->CallObjectMethod(task_obj, java_notify_error, j_error_msg);
                     local_env->DeleteLocalRef(j_error_msg);
                 }
             }
             else {
-                realm->close();
                 local_env->CallVoidMethod(task_obj, java_notify_realm_ready);
             }
+            local_env->DeleteGlobalRef(task_obj);
         });
         return reinterpret_cast<jlong>(&(*task));
     }